fix wrongful relative includes in tsconfig.json
For some reason specific Components were included in the tsconfig.json
with a relative path. This should have caused problems for anyone who
has not checked out `utopia-ui` in the coresponding location.
This was discovered due to the fact that we cannot include components
directly anymore since the include paths have project specific path
aliases