refactor(source): export point type (#138)
* bind the local utopia-ui lib

To ensure the examples still work with the latest utopia-ui, it is now
bound to the current code and does no longer depend on publishing the
library first

* replaced number by uuid as item ids in examples

* remove link mechanic from workflow

* also build utopia-ui

* fix example2

for some reason this stopped working on maaster.

* export geojson point to remove this external dependency

By exporting this type another package is not needed. This is especially
important, since there are different Point definitions which are
partially incompatible.

---------

Co-authored-by: Anton Tranelis <31516529+antontranelis@users.noreply.github.com>
Co-authored-by: Anton Tranelis <mail@antontranelis.de>